requirements.txt
Flask==3.1.3
Werkzeug==3.1.6
Jinja2==3.1.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
